chore: move entity into client/entity
其为 client 运行时产生的结构, 理应放置其下
