+++ client/entity/friend.go
