+++ client/entity/group.go
